Renaming views.

+++ src/main/java/pl/edu/pb/wi/model/Component.java
@@ -1,5 +1,5 @@
 package pl.edu.pb.wi.model;
 
 public interface Component {
-    Integer getPoint();
+    double getPoint();
 }
